chore: move getAccount... functions to account service
also renaming getAccountAndEnvironment... to getAccountContext...
in preparation of also returning the Plan
