+++ packages/shared/lib/services/environment.service.ts
@@ -2,7 +2,6 @@ import * as uuid from 'uuid';
 import * as z from 'zod';
 
 import db from '@nangohq/database';
-import { isCloud } from '@nangohq/utils';
 
 import { PROD_ENVIRONMENT_NAME } from '../constants.js';
 import { configService, externalWebhookService, getGlobalOAuthCallbackUrl } from '../index.js';
@@ -11,13 +10,11 @@ import encryptionManager, { pbkdf2 } from '../utils/encryption.manager.js';
 import errorManager, { ErrorSourceEnum } from '../utils/error.manager.js';
 
 import type { Orchestrator } from '../index.js';
-import type { DBEnvironment, DBEnvironmentVariable, DBTeam, SdkLogger } from '@nangohq/types';
+import type { DBEnvironment, DBEnvironmentVariable, SdkLogger } from '@nangohq/types';
 
 const TABLE = '_nango_environments';
 
 export const defaultEnvironments = [PROD_ENVIRONMENT_NAME, 'dev'];
-
-const hashLocalCache = new Map<string, string>();
 
 class EnvironmentService {
     async getEnvironmentsByAccountId(account_id: number): Promise<Pick<DBEnvironment, 'id' | 'name'>[]> {
@@ -42,143 +39,6 @@ class EnvironmentService {
 
             return [];
         }
-    }
-
-    async getAccountAndEnvironmentBySecretKey(secretKey: string): Promise<{ account: DBTeam; environment: DBEnvironment } | null> {
-        if (!isCloud) {
-            const environmentVariables = Object.keys(process.env).filter((key) => key.startsWith('NANGO_SECRET_KEY_'));
-            if (environmentVariables.length > 0) {
-                for (const environmentVariable of environmentVariables) {
-                    const envSecretKey = process.env[environmentVariable] as string;
-
-                    if (envSecretKey !== secretKey) {
-                        continue;
-                    }
-
-                    const envName = environmentVariable.replace('NANGO_SECRET_KEY_', '').toLowerCase();
-                    // This key is set dynamically and does not exist in database
-                    const env = await db.knex
-                        .select<Pick<DBEnvironment, 'account_id'>>('account_id')
-                        .from<DBEnvironment>(TABLE)
-                        .where({ name: envName, deleted: false })
-                        .first();
-
-                    if (!env) {
-                        return null;
-                    }
-
-                    return this.getAccountAndEnvironment({ accountId: env.account_id, envName });
-                }
-            }
-        }
-
-        return this.getAccountAndEnvironment({ secretKey });
-    }
-
-    async getAccountIdFromEnvironment(environment_id: number): Promise<number | null> {
-        const result = await db.knex.select('account_id').from<DBEnvironment>(TABLE).where({ id: environment_id, deleted: false });
-
-        if (result == null || result.length == 0 || result[0] == null) {
-            return null;
-        }
-
-        return result[0].account_id;
-    }
-
-    async getAccountFromEnvironment(environment_id: number): Promise<DBTeam | null> {
-        const result = await db.knex
-            .select<DBTeam>('_nango_accounts.*')
-            .from(TABLE)
-            .join('_nango_accounts', '_nango_accounts.id', '_nango_environments.account_id')
-            .where('_nango_environments.id', environment_id)
-            .first();
-
-        return result || null;
-    }
-
-    async getAccountAndEnvironmentByPublicKey(publicKey: string): Promise<{ account: DBTeam; environment: DBEnvironment } | null> {
-        if (!isCloud) {
-            const environmentVariables = Object.keys(process.env).filter((key) => key.startsWith('NANGO_PUBLIC_KEY_'));
-            if (environmentVariables.length > 0) {
-                for (const environmentVariable of environmentVariables) {
-                    const envPublicKey = process.env[environmentVariable] as string;
-
-                    if (envPublicKey !== publicKey) {
-                        continue;
-                    }
-                    const envName = environmentVariable.replace('NANGO_PUBLIC_KEY_', '').toLowerCase();
-                    // This key is set dynamically and does not exist in database
-                    const env = await db.knex
-                        .select<Pick<DBEnvironment, 'account_id'>>('account_id')
-                        .from<DBEnvironment>(TABLE)
-                        .where({ name: envName, deleted: false })
-                        .first();
-                    if (!env) {
-                        return null;
-                    }
-
-                    return this.getAccountAndEnvironment({ accountId: env.account_id, envName });
-                }
-            }
-        }
-
-        return this.getAccountAndEnvironment({ publicKey });
-    }
-
-    async getAccountAndEnvironment(
-        // TODO: fix this union type that is not discriminated
-        opts:
-            | { publicKey: string }
-            | { secretKey: string }
-            | { accountId: number; envName: string }
-            | { environmentId: number }
-            | { environmentUuid: string }
-            | { accountUuid: string; envName: string }
-    ): Promise<{ account: DBTeam; environment: DBEnvironment } | null> {
-        const q = db.readOnly
-            .select<{
-                account: DBTeam;
-                environment: DBEnvironment;
-            }>(db.knex.raw('row_to_json(_nango_environments.*) as environment'), db.knex.raw('row_to_json(_nango_accounts.*) as account'))
-            .from<DBEnvironment>(TABLE)
-            .join('_nango_accounts', '_nango_accounts.id', '_nango_environments.account_id')
-            .where('_nango_environments.deleted', false)
-            .first();
-
-        let hash: string | undefined;
-        if ('secretKey' in opts) {
-            // Hashing is slow by design so it's very slow to recompute this hash all the time
-            // We keep the hash in-memory to not compromise on security if the db leak
-            hash = hashLocalCache.get(opts.secretKey) || (await hashSecretKey(opts.secretKey));
-            q.where('secret_key_hashed', hash);
-        } else if ('publicKey' in opts) {
-            q.where('_nango_environments.public_key', opts.publicKey);
-        } else if ('environmentUuid' in opts) {
-            q.where('_nango_environments.uuid', opts.environmentUuid);
-        } else if ('accountUuid' in opts) {
-            q.where('_nango_accounts.uuid', opts.accountUuid).where('_nango_environments.name', opts.envName);
-        } else if ('accountId' in opts) {
-            q.where('_nango_environments.account_id', opts.accountId).where('_nango_environments.name', opts.envName);
-        } else if ('environmentId' in opts) {
-            q.where('_nango_environments.id', opts.environmentId);
-        } else {
-            return null;
-        }
-
-        const res = await q;
-        if (!res) {
-            return null;
-        }
-
-        if (hash && 'secretKey' in opts) {
-            // Store only successful attempt to not pollute the memory
-            hashLocalCache.set(opts.secretKey, hash);
-        }
-        return {
-            // Getting data with row_to_json breaks the automatic string to date parser
-            account: { ...res.account, created_at: new Date(res.account.created_at), updated_at: new Date(res.account.updated_at) },
-            environment: encryptionManager.decryptEnvironment(res.environment)
-        };
     }
 
     async getById(id: number): Promise<DBEnvironment | null> {
